Make the package lookups case-insensitive.
This changes the signature of the find_best_match() call, which now
returns a (name, version) tuple, which provides the correct casing of
the package name.

This matches the equivalent of real PyPI lookups, where when you search
for django<1.4, you get returned Django==1.3.3 (note the casing).  Also,
when asked for django_pipeline, you get django-pipeline==1.2.17.