PyPIRepository: keep old references to build_dir/source_dir
This is a bit hackish, but is required for the `source_dir` of a
requirement to be available in the end for `format_requirement`.

This might not be required when not caching the prepared requirements /
dependencies, but that is a) slower and b) would require other fixes to
move/adjust/reset the `source_dir` when the build dir gets pruned.