Comment out breaking check
This check currently breaks if the source directory is removed before it can run. It would still be good to have the check, but we first have to find out when/where the directory is removed and either prevent that or move the check to a location where we're sure the directory still exists.