Check we have a compatible pip version *before* other imports
If running an out-of-date version of pip, it's possible to hit an
ImportError while doing package-internal imports.  This means we
never hit the compatibility check, and a nasty traceback is
printed to the console.  See #428 for one such example.

This patch runs the compatibility checks as soon as we import pip,
which minimises the risk of an unhelpful error message.