Try fixing the cram test
At first I was not able to run the test locally on my mac at all.
Replacing the `alias` by a `function` fixed this. At this point I
was seeing the same problem as on Travis CI, where the
`python-dateutil` package is not being updated. Passing the
arguments of the function to the internal `python -m pip_review`
using `$*` fixed this. Hopefully, this fixes it on Travis as well.