Remove special casing of pip specifier attribute.
The use of req.specs in the resolver causes the resolver to not
stabilize.

This happens on long lists of subdependency. The old conditionnal was
introducing discrepancies in the comparison function, under pip < 8.1.2.

This change removes the pip version dependant check and
unconditionnally uses req.specifier. Now that pip-tools only supports
pip > 8.0, and that this attribute is present in pip
>= 8.0, the special case is not needed.