Use the same workaround for 'pip freeze' in get_installed_pkgs
This is necessary for consistency. However, the change revealed a
strange quirk in Pip<6.0 under Python>=2.7, which in turn breaks the
last test. The change to tests/review.t mitigates this. Please see
below for further discussion.

Pip<6.0 (i.e., Pip==1.5.6) behaves differently when invoked as
`python -m pip freeze` than when invoked as just `pip freeze`. In the
first case, it includes pip_review as an editable package from git.
Editable packages are skipped, but still cause the `all_ok` variable
to turn to `False`, so pip-review does not output anything while the
expected output is "Everything up-to-date".

This might be a problem that you can only run into when testing under
multiple versions of Python locally using Tox.