Improve setup.py parsing
Since requirements.txt should hold the reqs for the whole system, we should read from `setup_requires` _and_ `tests_require` when making the list of requirements.

Also avoid erroring out in an ugly if the user doesn't happen to have any requirements. Now, we throw a meaningful error if that's the case.