Fixed bug breaking pip-sync on Python 3, raising TypeError
Fixes #569
Moved and renamed `resolver._dep_key` to `utils.key_from_ireq` and use it
for the `InstallRequirement`s sort in `sync.sync`.

In the future, we might want to change calls like
`key_from_req(ireq.req)` to `key_from_ireq(ireq)`, but I'll keep the
change to a minimum here for this bugfix.