ci: codecov: remove flags, add OS name to job name
While pip-tools is not (yet?) that much affected by timeouts on
codecov.io's backend, it is certainly rather slow in general.

As it turned out with pytest (where it was constantly timing out), this
is due to the usage of flags.

Therefore this removes flags for now, but adds TRAVIS_OS_NAME to the
upload's name on Travis (which means that just TOXENV is coming from
AppVeyor).

On Travis, this also moves the after_success section to the end of the
config (where you would expect it).

It also only installs codecov/coveralls when necessary.