Add -U option to install_command
There was an issue with PIP=latest because of cache. To fix it added
to pip upgrade option to make sure that all packages
(without explicit version) upgraded to a latest version.

Also print explicitly pip version for debugging. Now you know
exactly which version of pip was installed on PIP=latest