Always place annotations on a new line
This makes the annotation output more consistent. No longer place
annotations inline when hashes aren't generated. Now, always place
annotations on a new line. When the annotation comment is very long,
this helps make the result more readable as it is less likely to
continue far off to the right.