Use context manager to restore environment variables in resolve()
In the event of an unexpected exception, the system's environment
variable PIP_EXISTS_ACTION should still be restored to its previous
state.

This context manager already exists in the upstream pip library. Can
reuse it to avoid duplication.

Unlike the previous implementation, if the PIP_EXISTS_ACTION environment
variable existed prior to executing pip-tools, that state will be
restored.