Remove unused .fussyfox.yml
At some point, FussyFox was turned off. It no longer runs on PRs. So
this file goes unused. As it is unused, it only adds clutter to the top
level directory which is already full of other configuration files.

Regardless, I'm not sure this is needed any longer. After the move to
GitHub Actions, CI is much faster and the lint targets run in parallel
to the tests. I think whatever benefit it provided is now builtin.