Remove unreachable code in PyPIRepository.get_hashes()
get_hashes() is only called after the dependencies have been resolved.
After they have been resolved, requirements are either links or pinned
requirements.

If the requirement is a link, the "if ireq.link" block is executed and
the function returns early. If the requirement is not a link, then it
has been resolved to an exact version. Either way, the exception will
never be raised.

This line is untested and so frequently appears as uncovered during
coverage analysis, erroneously marking the PR with a big red ❌ when the
PR is otherwise green.