write default index in "--extra-index-url"
If the default index is provided as `--extra-index-url` it can not
safely be dropped from the written output since this changes behaviour
when the output is synced.

While jazzband/pip-tools#756 Explicitly introduced tests for the
opposite this seams to have been a mistake since otherwise the following
lines comment about making an exception would be nonsensical.