Add release drafter github action
This change should help preparing list of release notes for future
releases. When a maintainer will want to make a release he would
be able to tune the draft release notes and copy them to the
changelog.

The biggest advantage of this change is that the creation of the
draft is automated and it already uses the desired format, with
title link to PR and mention of author, reducing the maintenance
chore when preparing for a new release.