Add optional from_dir param to parse_requirements
If passed, from_dir will be used instead of CWD to interpret relative paths.

Special care is taken to guarantee _was_relative in the yielded ireqs,
if originally relative.

When calling format_requirement,
_was_relative's presence now determines what's passed as from_dir:
either CWD or None (None leads to an absolute path output)

For ths to work we also re-assign the _was_relative attr
to ireqs generated from candidates in the backtracking resolver.