Add flag to sync: --read-relative-to-input
With test_sync_relative_path

If the reqs in the output were written relative to that output file
(pip-compile --write-relative-to-output), then it's wise for pip-sync
to interpret reqs it finds in there from the location of the output file.

This is now possible with pip-sync --read-relative-to-input.

To be clear: pip-sync's "input" is pip-compile's "output."