🐛 Use requested constraints for build backend
This fixes a bug that build deps compilation would get the latest
version of an unconstrained build requirements list, not taking into
account the restricted/requested one.

The regression test is implemented against `setuptools < 70.1.0` which
is known to inject a dependency on `wheel` (newer `setuptools` vendor
it). The idea is that `pyproject.toml` does not have an upper bound for
`setuptools` but the CLI arg does. And when this works correctly, the
`wheel` entry will be included into the resolved output.

Cleaning up `PIP_CONSTRAINT` is implemented manually due to the corner
case of a permission error on Windows when accessing a file that we
hold a file descriptor to from another subprocess[[1]]. It can be
further simplified once the lowest Python version `pip-tools` supports
is 3.12 by replacing `delete=False` with `delete_on_close=False` in the
`tempfile.NamedTemporaryFile()` context manager initializer.

[1]: https://docs.python.org/3/library/tempfile.html#tempfile.NamedTemporaryFile

Co-authored-by: Sviatoslav Sydorenko <webknjaz@redhat.com>