Add tests for [dependency-groups]
And apply fixes and changes as required.
Tests include a simple positive test case and a number of error
checking cases for different scenarios.

Minor fixes to output are needed and `--group` is treated as an
alternative to having a source file when determining whether or not
sufficient input options were given.