Conditionally rewrite abspath 'comes_from' data
In order to support newer `pip` versions, `pip-compile` needs to
handle the fact that the `comes_from` field on a parsed requirement
may be rewritten from a relative path to an absolute one.

The change in `pip` occurred in version 24.3. `comes_from` is
populated with absolute paths.

`pip-compile` targets usage in which a relative path is supplied as
input and ought to be preserved in the annotations for its outputs. As
such, even though the relative path has been translated to an absolute
one, the original (relative) path should be rendered.

Because the data coming back from `pip` are ambiguous to some degree,
as a heuristic, this change will use whether or not the original path
was absolute to determine behavior. Therefore, the following files
will result in correct behavior:

    # reqs.in
    -r other-reqs.in
    # other-reqs.in
    some-pkg

`some-pkg` came from `-r other-reqs.in`.

By contrast, if the CWD is `/opt`, and `-r reqs.in` is used, these
data will be *incorrectly* rendered with relative paths:

    # reqs.in
    -r /opt/other-reqs.in
    # /opt/other-reqs.in
    some-pkg

`some-pkg` came from `-r /opt/other-reqs.in`, but will be annotated
with `-r other-reqs.in`

Note that the relative path which is used is, in this scenario,
correct, but not reflective of the inputs to the program.

However, it is possible, if the absolute paths are really wanted in
such a case, to ask for them using `pip-compile -r /opt/reqs.in`.