Introduce initial Towncrier config
Modeled on a combination of `cheroot/towncrier.toml` and
`attrs/changelog.d/towncrier_template.md.jinja`, these files are an
initial configuration with minimal complexity.

For example, the `cheroot` code block which looks up issue numbers is
omitted for simplicity.

An initial basic `tox r -e changelog` environment is added here to let
us verify the config.

The contribution doc is modeled on the `attrs` one.