Fix rendering of the draft changelog
MyST does not properly evaluate `{only}` directives, and
experimentally always treats them as `true`. As a result, the draft
changelog is included even when a release build is set.

To fix, nest a layer of `eval-rst`, to drop into the sphinx rst parser
which handles these correctly, into the changelog. An explanatory
comment describes why we are doing this.

In order to support this usage, the changelog draft MyST directive is
broken out into a separate partial doc for inclusion. For now, it's
placed into `changelog.d/` to keep it clearly outside of the doc build
targets.

MyST flags the partial for starting at the wrong header level. This
appears to be a false positive for that check -- the header level is
intentionally starting at H2 for this partial doc. It does not appear
to be possible to turn off the warning with specificity or to use
MyST directives to clarify our intent. `:start-after:` for the include
directive, with a dummy header, is similarly ineffective, as MyST
seems to be warning when it loads the partial doc.
For now, we can simply suppress the warning altogether.

In order to facilitate testing -- and potentially unifying our RTD
build with our tox config in the future -- RTD environment variables
are now declared in passenv for the `build-docs` environment.