Link to genuine pull requests from README examples
The migration to towncrier in #2203 included a couple of
synthetic change note examples.

Now that towncrier is in use and some changes that use it
have landed in pip-tools, we can update the examples so that
they hyperlink to genuine pull requests.

This is intended to aid contributors by providing them with
more accurate, representative examples to crib from.