Pin pip-tools' dependencies for "build-docs"
In order to get consistent doc builds, we need the dependencies of
`pip-tools` (in particular, right now we need `pip`) pinned to
versions where the build works.

In order to provide a `requirements.txt` file with these pins, we add
`.` to `docs/requirements.in`. This requires a couple of tweaks:

- Use `-e .` so that `pip-compile` preserves it as a relative path
- Remove the `changedir` tox directive, so that doc building and
  pinning happens from the tox ini dir (repo root)

Additionally, `pip` itself is downgraded from 25.2 (where we have the
doc build issue today) to 25.1 .

Other updates to `docs/requirements.txt` occur as side-effects of this
change. New packages are added and paths are updated.