Make 'pyproject.toml' absolute for use as URI
When statically parsing a pyproject.toml , we can encounter
self-referential package usage. For example, an extra which names
another extra on the same package.

The source for these requirements is the current package, but as a
concrete requirement, not an abstract one. To make it concrete, we use
a URI dependency link pointed at the package directory as a file-URI.

When the input path to `pyproject.toml` is relative, it needs to be
made absolute for use as a URI. Failing to do so results in errors.

Because the `pip-compile` output of an such an extra depends
concretely on the specific version of the package found in-tree, and
not abstractly on the package name (without the URI), tests confirm
that we see the rendered package directory as a file URI in
`pip-compile` output.