Bugfix: path normalization applied to URLs
The `-r` and `-c` path normalization logic did not check if the
requested requirements or constraint files were remote files being
loaded, e.g., via HTTPS. As a result, a usage like
`-c https://example.com/constraints.txt` incorrectly got converted to
a path and normalized, stripping one of the two slashes which
separates the scheme from the netloc.

A regression test is added which reproduces the bug, and the gap in
detection is closed by checking explicitly if inputs can parse as
URIs. Any given URI (including file URIs) will be exempted from the
path rewrites.

The check is implemented by calling `urllib.parse.urlparse()` and
checking the `scheme` attribute of the parse result.