📦 Build release artifacts in CI
This is a narrowly scoped initial integration of `reusable-tox.yml`
into the project's CI/CD infra. It makes use of the following external
action and reusable workflow repositories, some are optional:
* `codecov/codecov-action` (transitively, via `tox-dev/workflow`)
* `irongut/CodeCoverageSummary` (transitively, via `tox-dev/workflow`)
* `re-actors/cache-python-deps` « (directly; and also transitively,
  via `tox-dev/workflow`)
* `re-actors/checkout-python-sdist` «
  (transitively, via `tox-dev/workflow`)
* `test-summary/action` (transitively, via `tox-dev/workflow`)
* `tox-dev/workflow` « (directly)

The marked projects are under my control.

This patch plugs the main CI workflow as reusable into the release
one. The CI workflow gets two new jobs — pre-setup that computes a
number of variables, and build that makes the artifacts and stores
them for later consumption by the PyPI upload job. The build job
also double-checks that the artifacts are created with expected
names.

The invocation of `pypa/build` moved to tox so that it's possible to
run the same process locally.

In follow-up PRs, these dists will also be used in testing.